MU2_BGFX sprint 5, sent back: an invariant that caught one of fifty-two, and two rates that were not floats
QA reviewed both halves of sprint 5 and returned SENT BACK with seven findings. It ran what the
sprint file claimed rather than reading it -- the hunt at four seeds, with steps, headless-only,
on Noria, at 50 000 ticks, and a second build directory at -O0 replaying byte-identical to
Release on both maps, which is a stronger statement about determinism than this sprint had made
for itself. All seven are answered.

The two that were defects rather than wording:

  * **"No blow landed on the dead" was very nearly vacuous.** It looked for a death earlier in
    THE SAME tick's happenings, and happenings are cleared every tick -- so it could only catch
    a swing at something that had died that very tick. QA removed the guard in the sim on
    purpose: 52 swings landed on corpses and the check reported 1. Findings now carries who is
    dead between ticks; broken the same way again it reports 330. This is the one that mattered,
    because the sprint file says tests/ replaces Instrument, Audit and the bot.
  * **The attack rate and the defence rate were truncated to int and OpenMU keeps them float.**
    Half of QA's finding: read at the source, OpenMU truncates in two of the three places and
    not in the third. `GetHitChanceTo` declares both rates `float` and divides them as floats
    (AttackableExtensions.cs:694-716), and `Overrates` compares them as floats (:728-731) -- so
    a knight's 6.667 of defence rate was 6 here and his hit chance 0.900 where 0.890 is right.
    But the defence IS `(int)((attribute + bonus) * decrement)` at :92, and the damage band comes
    out of `GetBaseDmg` as `out int` (:837-849), so those two were already faithful and carrying
    them as floats would have been wrong in the other direction. Fighter now holds two floats
    and three integers, and each one names the cast of OpenMU's it is.

And five smaller ones: the 65-and-over experience term was floored where OpenMU's division is
float (unreachable in this content, which is exactly why the branch is kept at all); the drawn
figures were a whole tick further behind than their own comment claimed, because `remember()`
ran before `step()` rather than after; "a step allocates nothing else" was untrue of a route
growing into a body's own vector, so every body reserves its tiles at spawn and the header now
states the bound it actually has; the sprint file quoted MU2's `&& attackRate > 0` guard as
OpenMU's trace, which the code itself had always marked honestly; and "Lorencia holds 250
monsters" sat in a paragraph listing nine numbers that sum to 290.

The rules changed, so the seeded log's bytes changed with them: seed 1, level 4, 10 000 ticks is
94 466 bytes, fingerprint dba488f9089a11a4, identical twice; seed 7 at 50 000 identical twice;
Noria identical and clean; step cost unchanged at 0.0009 ms on Lorencia and 0.0034 on Noria.

Not ours, and left alone: the `present` account reads 2.785 ms against its 0.500 budget in every
window run with and without --play, which predates this sprint; and the Dark Knight's held
weapon sits as a flat plate at hip height, which is sprint 4's attachment.

+++ src/sim/rules.h
@@ -21,8 +21,17 @@ constexpr int kPointsPerLevel = 5;
 // Everything a blow reads off either side of it.
 struct Fighter {
     int level = 1;
-    int attackRate = 0;
-    int defenseRate = 0;
+    // The two rates are FLOATS and the three below them are integers, and that is not
+    // carelessness -- it is where OpenMU's own casts fall. `GetHitChanceTo`
+    // (AttackableExtensions.cs:694-716) declares both rates `float` and divides them as
+    // floats, and `Overrates` (:728-731) compares them as floats; a knight's 6.667 of defense
+    // rate truncated to 6 gives a hit chance of 0.900 where the original gives 0.890.
+    // Meanwhile the defence is `(int)((attribute + bonus) * decrement)` at :92 and the damage
+    // band comes out of `GetBaseDmg` already cast (`out int`, :837-849) -- so those three are
+    // truncated in the original too, at exactly these points, and carrying them as floats here
+    // would be as wrong in the other direction.
+    float attackRate = 0.0f;
+    float defenseRate = 0.0f;
     int defense = 0;
     int minimumDamage = 0;
     int maximumDamage = 0;
@@ -54,7 +63,7 @@ struct Blow {
 // left to be found: OpenMU divides without it, which is a division by zero for an attacker
 // with no attack rate at all. Nothing in 0.75's data has one. The guard changes no outcome
 // this content can produce and removes an undefined one it cannot.
-double hitChance(int attackRate, int defenseRate);
+double hitChance(float attackRate, float defenseRate);
 
 // One swing, in OpenMU's own order, and the order is the behaviour rather than the
 // presentation. AttackableExtensions.cs:69-225, physical arm, with every branch 0.75 cannot
